napi

Module bindgen_prelude

Re-exports

pub use crate::error;

pub use crate::JsError;

pub use crate::Property;

pub use crate::PropertyAttributes;

pub use crate::Result;

pub use crate::Status;

pub use crate::Task;

pub use crate::ValueType;

pub use crate::Env;

pub use crate::JsFunction;

pub use crate::sys;

Modules

iterator

Structs

AbortSignal

https://developer.mozilla.org/zh-CN/docs/Web/API/AbortController

Array

AsyncTask

Buffer

Zero copy u8 vector shared between rust and napi. It’s designed to be used in async context, so it contains overhead to ensure the underlying data is not dropped. For non-async context, use BufferRef instead.

BufferSlice

Zero copy buffer slice shared between Rust and Node.js. It can only be used in non-async context and the lifetime is bound to the fn closure. If you want to use Node.js Buffer in async context or want to extend the lifetime, use Buffer instead.

ClassInstance

Error

Represent JsError. Return this Error in js_function, napi-rs will throw it as JsError for you. If you want throw it as TypeError or RangeError, you can use JsTypeError/JsRangeError::from(Error).throw_into(env)

ExtendedErrorInfo

External

Float32Array

Float64Array

Function

A JavaScript function. It can only live in the scope of a function call. If you want to use it outside the scope of a function call, you can turn it into a reference. By calling the create_ref method.

FunctionRef

A reference to a JavaScript function. It can be used to outlive the scope of the function.

Int8Array

Int16Array

Int32Array

JsError

JsRangeError

JsTypeError

Null

Reference

Experimental feature

SharedReference

Experimental feature

Symbol

Uint8Array

Uint8ClampedArray

Uint8ClampedSlice

Zero copy Uint8ClampedArray slice shared between Rust and Node.js. It can only be used in non-async context and the lifetime is bound to the fn closure. If you want to use Node.js Uint8ClampedArray in async context or want to extend the lifetime, use Uint8ClampedArray instead.

Uint16Array

Uint32Array

Unknown

Utf16String

WeakReference

Enums

Either

Either3

Either4

Either5

Either6

Either7

Either8

Either9

Either10

Either11

Either12

Either13

Either14

Either15

Either16

Either17

Either18

Either19

Either20

Either21

Either22

Either23

Either24

Either25

Either26

TypedArrayType

Traits

FromNapiMutRef

FromNapiRef

FromNapiValue

Generator

Implement a Iterator for the JavaScript Class. This feature is an experimental feature and is not yet stable.

JsValuesTupleIntoVec

ObjectFinalize

ToNapiValue

TypeName

ValidateNapiValue

Functions

get_c_callback

Get C Callback from defined Rust fn

get_js_function

Get JsFunction from defined Rust fn

napi_register_module_v1⚠

Register the n-api module exports.

within_runtime_if_available

If the feature tokio_rt has been enabled this will enter the runtime context and then call the provided closure. Otherwise it will just call the provided closure.

Type Aliases

ExportRegisterCallback

ModuleExportsCallback

Object

Result

This

Undefined

Attribute Macros

ctor

Marks a function or static variable as a library/executable constructor. This uses OS-specific linker sections to call a specific function at load time.