napi::bindgen_prelude

Function napi_register_module_v1

#[no_mangle]
pub unsafe extern "C" fn napi_register_module_v1(
    env: napi_env,
    exports: napi_value,
) -> napi_value

Register the n-api module exports.

Safety

This method is meant to be called by Node.js while importing the n-api module. Only call this method if the current module is not imported by a node-like runtime.

Arguments env and exports must not be null.