Struct napi::bindgen_prelude::BigInt

pub struct BigInt {
    pub sign_bit: bool,
    pub words: Vec<u64>,
}

https://nodejs.org/api/n-api.html#napi_create_bigint_words The resulting BigInt is calculated as: (–1)^sign_bit (words[0] × (2^64)^0 + words[1] × (2^64)^1 + …)

Fields
sign_bit: bool

true for negative numbers

words: Vec<u64>
Implementations

impl BigInt

pub fn get_u64(&self) -> (bool, u64, bool)

(signed, value, lossless) get the first word of the BigInt as u64 return true in the last element of tuple if the value is lossless or the value is truncated

pub fn get_i64(&self) -> (i64, bool)

(value, lossless) get the first word of the BigInt as i64 return true if the value is lossless or the value is truncated

pub fn get_i128(&self) -> (i128, bool)

(value, lossless) get the first two words of the BigInt as i128 return true if the value is lossless or the value is truncated

pub fn get_u128(&self) -> (bool, u128, bool)

(signed, value, lossless) get the first two words of the BigInt as u128 return true if the value is lossless or the value is truncated