napi::bindgen_prelude

Struct External

pub struct External<T: 'static> {
    pub adjusted_size: i64,
    /* private fields */
}

Fields
adjusted_size: i64
Implementations

impl<T: 'static> External<T>

pub fn new(value: T) -> Self

pub unsafe fn inner_from_raw_mut( unknown_tagged_object: *mut c_void, ) -> Option<&'static mut T>

Turn a raw pointer (from napi) pointing to the inner *mut TaggedObject<T> into a reference inner object.

Safety

The unknown_tagged_object raw pointer must point to an TaggedObject<T> struct, which is essentially the pointer which napi-rs returns to the NAPI api.

Additionally, you must ensure that there are no other live mutable references to the T for the duration of the lifetime of the returned mutable reference.

pub unsafe fn inner_from_raw( unknown_tagged_object: *mut c_void, ) -> Option<&'static T>

Turn a raw pointer (from napi) pointing to the inner *mut TaggedObject<T> into a reference inner object.

Safety

The unknown_tagged_object raw pointer must point to an TaggedObject<T> struct, which is essentially the pointer which napi-rs returns to the NAPI api.

Additionally, you must ensure that there are no other live mutable references to the T for the duration of the lifetime of the returned immutable reference.

pub fn new_with_size_hint(value: T, size_hint: usize) -> Self

size_hint is a value to tell Node.js GC how much memory is used by this External object.

If getting the exact size_hint is difficult, you can provide an approximate value, itโs only effect to the GC.

If your External object is not effect to GC, you can use External::new instead.