napi::bindgen_prelude

Struct Function

pub struct Function<'scope, Args: JsValuesTupleIntoVec, Return: FromNapiValue> { /* private fields */ }

A JavaScript function. It can only live in the scope of a function call. If you want to use it outside the scope of a function call, you can turn it into a reference. By calling the create_ref method.

Implementations

impl<'scope, Args: JsValuesTupleIntoVec, Return: FromNapiValue> Function<'scope, Args, Return>

pub fn call(&self, args: Args) -> Result<Return>

Call the JavaScript function. this in the JavaScript function will be undefined. If you want to specify this, you can use the apply method.

pub fn apply<Context: ToNapiValue>( &self, this: Context, args: Args, ) -> Result<Return>

Call the JavaScript function. this in the JavaScript function will be the provided this.

pub fn create_ref(&self) -> Result<FunctionRef<Args, Return>>

Create a reference to the JavaScript function.