Struct napi::JsObject

pub struct JsObject(_);

Implementations

impl JsObject

pub fn get<K: AsRef<str>, V: FromNapiValue>( &self, field: K ) -> Result<Option<V>>

pub fn set<K: AsRef<str>, V: ToNapiValue>( &mut self, field: K, val: V ) -> Result<()>

pub fn keys(obj: &Object) -> Result<Vec<String>>

impl JsObject

pub fn add_finalizer<T, Hint, F>( &mut self, native: T, finalize_hint: Hint, finalize_cb: F ) -> Result<()>where T: 'static, Hint: 'static, F: FnOnce(FinalizeContext<T, Hint>),

impl JsObject

pub fn into_unknown(self) -> JsUnknown

pub fn coerce_to_bool(self) -> Result<JsBoolean>

pub fn coerce_to_number(self) -> Result<JsNumber>

pub fn coerce_to_string(self) -> Result<JsString>

pub fn coerce_to_object(self) -> Result<JsObject>

pub fn is_date(&self) -> Result<bool>

pub fn is_promise(&self) -> Result<bool>

pub fn is_error(&self) -> Result<bool>

pub fn is_typedarray(&self) -> Result<bool>

pub fn is_dataview(&self) -> Result<bool>

pub fn is_array(&self) -> Result<bool>

pub fn is_buffer(&self) -> Result<bool>

pub fn instanceof<Constructor>(&self, constructor: Constructor) -> Result<bool>where Constructor: NapiRaw,

impl JsObject

pub fn set_property<K, V>(&mut self, key: K, value: V) -> Result<()>where K: NapiRaw, V: NapiRaw,

pub fn get_property<K, T>(&self, key: K) -> Result<T>where K: NapiRaw, T: NapiValue,

pub fn get_property_unchecked<K, T>(&self, key: K) -> Result<T>where K: NapiRaw, T: NapiValue,

pub fn set_named_property<T>(&mut self, name: &str, value: T) -> Result<()>where T: ToNapiValue,

pub fn create_named_method( &mut self, name: &str, function: Callback ) -> Result<()>

pub fn get_named_property<T>(&self, name: &str) -> Result<T>where T: FromNapiValue,

pub fn get_named_property_unchecked<T>(&self, name: &str) -> Result<T>where T: FromNapiValue,

pub fn has_named_property(&self, name: &str) -> Result<bool>

pub fn delete_property<S>(&mut self, name: S) -> Result<bool>where S: NapiRaw,

pub fn delete_named_property(&mut self, name: &str) -> Result<bool>

pub fn has_own_property(&self, key: &str) -> Result<bool>

pub fn has_own_property_js<K>(&self, key: K) -> Result<bool>where K: NapiRaw,

pub fn has_property(&self, name: &str) -> Result<bool>

pub fn has_property_js<K>(&self, name: K) -> Result<bool>where K: NapiRaw,

pub fn get_property_names(&self) -> Result<JsObject>

pub fn get_all_property_names( &self, mode: KeyCollectionMode, filter: KeyFilter, conversion: KeyConversion ) -> Result<JsObject>

https://nodejs.org/api/n-api.html#n_api_napi_get_all_property_names return Array of property names

pub fn get_prototype<T>(&self) -> Result<T>where T: NapiValue,

This returns the equivalent of Object.getPrototypeOf (which is not the same as the function’s prototype property).

pub fn get_prototype_unchecked<T>(&self) -> Result<T>where T: NapiValue,

pub fn set_element<T>(&mut self, index: u32, value: T) -> Result<()>where T: NapiRaw,

pub fn has_element(&self, index: u32) -> Result<bool>

pub fn delete_element(&mut self, index: u32) -> Result<bool>

pub fn get_element<T>(&self, index: u32) -> Result<T>where T: NapiValue,

pub fn get_element_unchecked<T>(&self, index: u32) -> Result<T>where T: NapiValue,

pub fn define_properties(&mut self, properties: &[Property]) -> Result<()>

This method allows the efficient definition of multiple properties on a given object.

pub fn get_array_length(&self) -> Result<u32>

Perform is_array check before get the length if Object is not array, ArrayExpected error returned

pub fn get_array_length_unchecked(&self) -> Result<u32>

use this API if you can ensure this Object is Array

pub fn freeze(&mut self) -> Result<()>

pub fn seal(&mut self) -> Result<()>