Struct napi::JsTypedArray

pub struct JsTypedArray(_);

Implementations

impl JsTypedArray

pub fn into_value(self) -> Result<JsTypedArrayValue>

get TypeArray info https://nodejs.org/api/n-api.html#n_api_napi_get_typedarray_info

Warning: Use caution while using this API since the underlying data buffer is managed by the VM.

impl JsTypedArray

pub fn into_unknown(self) -> JsUnknown

pub fn coerce_to_number(self) -> Result<JsNumber>

pub fn coerce_to_string(self) -> Result<JsString>

pub fn coerce_to_object(self) -> Result<JsObject>

pub fn is_promise(&self) -> Result<bool>

pub fn is_error(&self) -> Result<bool>

pub fn is_typedarray(&self) -> Result<bool>

pub fn is_dataview(&self) -> Result<bool>

pub fn is_array(&self) -> Result<bool>

pub fn is_buffer(&self) -> Result<bool>

pub fn instanceof<Constructor>(&self, constructor: Constructor) -> Result<bool> where
Constructor: NapiRaw,

impl JsTypedArray

pub fn set_property<V>(&mut self, key: JsString, value: V) -> Result<()> where
V: NapiRaw,

pub fn get_property<K, T>(&self, key: K) -> Result<T> where
K: NapiRaw,
T: NapiValue,

pub fn get_property_unchecked<K, T>(&self, key: K) -> Result<T> where
K: NapiRaw,
T: NapiValue,

pub fn set_named_property<T>(&mut self, name: &str, value: T) -> Result<()> where
T: NapiRaw,

pub fn create_named_method(
&mut self,
name: &str,
function: Callback
) -> Result<()>

pub fn get_named_property<T>(&self, name: &str) -> Result<T> where
T: NapiValue,

pub fn get_named_property_unchecked<T>(&self, name: &str) -> Result<T> where
T: NapiValue,

pub fn has_named_property(&self, name: &str) -> Result<bool>

pub fn delete_property<S>(&mut self, name: S) -> Result<bool> where
S: NapiRaw,

pub fn delete_named_property(&mut self, name: &str) -> Result<bool>

pub fn has_own_property(&self, key: &str) -> Result<bool>

pub fn has_own_property_js<K>(&self, key: K) -> Result<bool> where
K: NapiRaw,

pub fn has_property(&self, name: &str) -> Result<bool>

pub fn has_property_js<K>(&self, name: K) -> Result<bool> where
K: NapiRaw,

pub fn get_property_names(&self) -> Result<JsObject>

pub fn get_prototype<T>(&self) -> Result<T> where
T: NapiValue,

This returns the equivalent of Object.getPrototypeOf (which is not the same as the function’s prototype property).

pub fn get_prototype_unchecked<T>(&self) -> Result<T> where
T: NapiValue,

pub fn set_element<T>(&mut self, index: u32, value: T) -> Result<()> where
T: NapiRaw,

pub fn has_element(&self, index: u32) -> Result<bool>

pub fn delete_element(&mut self, index: u32) -> Result<bool>

pub fn get_element<T>(&self, index: u32) -> Result<T> where
T: NapiValue,

pub fn get_element_unchecked<T>(&self, index: u32) -> Result<T> where
T: NapiValue,

pub fn define_properties(&mut self, properties: &[Property<'_>]) -> Result<()>

This method allows the efficient definition of multiple properties on a given object.

pub fn get_array_length(&self) -> Result<u32>

Perform is_array check before get the length if Object is not array, ArrayExpected error returned

pub fn get_array_length_unchecked(&self) -> Result<u32>

use this API if you can ensure this Object is Array