Module napi::bindgen_prelude

Re-exports

pub use crate::assert_type_of;

pub use crate::check_pending_exception;

pub use crate::check_status;

pub use crate::check_status_or_throw;

pub use crate::error;

pub use crate::error;

pub use crate::type_of;

pub use crate::JsError;

pub use crate::Property;

pub use crate::PropertyAttributes;

pub use crate::Result;

pub use crate::Status;

pub use crate::Task;

pub use crate::ValueType;

pub use crate::Env;

pub use crate::JsDate as Date;

pub use crate::JsFunction;

pub use crate::sys;

Modules

iterator

Structs

AbortSignal

https://developer.mozilla.org/zh-CN/docs/Web/API/AbortController

Array

AsyncTask

BigInt

https://nodejs.org/api/n-api.html#napi_create_bigint_words The resulting BigInt is calculated as: (–1)^sign_bit (words[0] × (2^64)^0 + words[1] × (2^64)^1 + …)

BigInt64Array

BigUint64Array

Buffer

Zero copy u8 vector shared between rust and napi. Auto reference the raw JavaScript value, and release it when dropped. So it is safe to use it in async fn, the &[u8] under the hood will not be dropped until the drop called. Clone will create a new Reference to the same underlying JavaScript Buffer.

CallbackInfo

ClassInstance

Error

Represent JsError. Return this Error in js_function, napi-rs will throw it as JsError for you. If you want throw it as TypeError or RangeError, you can use JsTypeError/JsRangeError::from(Error).throw_into(env)

ExtendedErrorInfo

External

Float32Array

Float64Array

Int8Array

Int16Array

Int32Array

JsError

JsRangeError

JsTypeError

Null

Promise

Reference

Experimental feature

SharedReference

Experimental feature

Symbol

Uint8Array

Uint8ClampedArray

Uint16Array

Uint32Array

Unknown

Utf16String

WeakReference

i64n

i64 is converted to Number

Enums

Either

Either3

Either4

Either5

Either6

Either7

Either8

Either9

Either10

Either11

Either12

Either13

Either14

Either15

Either16

Either17

Either18

Either19

Either20

Either21

Either22

Either23

Either24

Either25

Either26

TypedArrayType

Traits

FromNapiMutRef

FromNapiRef

FromNapiValue

Generator

Implement a Iterator for the JavaScript Class. This feature is an experimental feature and is not yet stable.

ObjectFinalize

ToNapiValue

TypeName

ValidateNapiValue

Functions

block_on

Runs a future to completion This is blocking, meaning that it pauses other execution until the future is complete, only use it when it is absolutely necessary, in other places use async functions instead.

execute_tokio_future

get_c_callback

Get C Callback from defined Rust fn

get_js_function

Get JsFunction from defined Rust fn

spawn

Spawns a future onto the Tokio runtime.

within_runtime_if_available

If the feature tokio_rt has been enabled this will enter the runtime context and then call the provided closure. Otherwise it will just call the provided closure.

Type Definitions

ExportRegisterCallback

ModuleExportsCallback

Object

Result

This

Undefined

Attribute Macros

ctor

Marks a function or static variable as a library/executable constructor. This uses OS-specific linker sections to call a specific function at load time.