Struct napi::Env

pub struct Env(_);

Env is used to represent a context that the underlying N-API implementation can use to persist VM-specific state.

Specifically, the same Env that was passed in when the initial native function was called must be passed to any subsequent nested N-API calls.

Caching the Env for the purpose of general reuse, and passing the Env between instances of the same addon running on different Worker threads is not allowed.

The Env becomes invalid when an instance of a native addon is unloaded.

Notification of this event is delivered through the callbacks given to Env::add_env_cleanup_hook and Env::set_instance_data.

Implementations

impl Env

pub unsafe fn from_raw(env: napi_env) -> Self

pub fn get_undefined(&self) -> Result<JsUndefined>

Get JsUndefined value

pub fn get_null(&self) -> Result<JsNull>

pub fn get_boolean(&self, value: bool) -> Result<JsBoolean>

pub fn create_int32(&self, int: i32) -> Result<JsNumber>

pub fn create_int64(&self, int: i64) -> Result<JsNumber>

pub fn create_uint32(&self, number: u32) -> Result<JsNumber>

pub fn create_double(&self, double: f64) -> Result<JsNumber>

pub fn create_bigint_from_i64(&self, value: i64) -> Result<JsBigint>

n_api_napi_create_bigint_int64

pub fn create_bigint_from_u64(&self, value: u64) -> Result<JsBigint>

pub fn create_bigint_from_i128(&self, value: i128) -> Result<JsBigint>

pub fn create_bigint_from_u128(&self, value: u128) -> Result<JsBigint>

pub fn create_bigint_from_words(
&self,
sign_bit: bool,
words: Vec<u64>
) -> Result<JsBigint>

n_api_napi_create_bigint_words

The resulting BigInt will be negative when sign_bit is true.

pub fn create_string(&self, s: &str) -> Result<JsString>

pub fn create_string_from_std(&self, s: String) -> Result<JsString>

pub fn create_string_from_vec_u8(&self, bytes: Vec<u8>) -> Result<JsString>

pub fn create_string_from_vec_i8(&self, bytes: Vec<i8>) -> Result<JsString>

pub fn create_string_utf16(&self, chars: &[u16]) -> Result<JsString>

pub fn create_string_latin1(&self, chars: &[u8]) -> Result<JsString>

pub fn create_symbol_from_js_string(
&self,
description: JsString
) -> Result<JsSymbol>

pub fn create_symbol(&self, description: Option<&str>) -> Result<JsSymbol>

pub fn create_object(&self) -> Result<JsObject>

pub fn create_array(&self) -> Result<JsObject>

pub fn create_array_with_length(&self, length: usize) -> Result<JsObject>

pub fn create_buffer(&self, length: usize) -> Result<JsBufferValue>

This API allocates a node::Buffer object. While this is still a fully-supported data structure, in most cases using a TypedArray will suffice.

pub fn create_buffer_with_data(&self, data: Vec<u8>) -> Result<JsBufferValue>

This API allocates a node::Buffer object and initializes it with data backed by the passed in buffer.

While this is still a fully-supported data structure, in most cases using a TypedArray will suffice.

pub unsafe fn create_buffer_with_borrowed_data<Hint, Finalize>(
&self,
data: *const u8,
length: usize,
hint: Hint,
finalize_callback: Finalize
) -> Result<JsBufferValue> where
Finalize: FnOnce(Hint, Env),

Safety

Mostly the same with create_buffer_with_data

Provided finalize_callback will be called when Buffer got dropped.

You can pass in noop_finalize if you have nothing to do in finalize phase.

pub fn adjust_external_memory(&mut self, size: i64) -> Result<i64>

This function gives V8 an indication of the amount of externally allocated memory that is kept alive by JavaScript objects (i.e. a JavaScript object that points to its own memory allocated by a native module).

Registering externally allocated memory will trigger global garbage collections more often than it would otherwise.

ATTENTION ⚠️, do not use this with create_buffer_with_data/create_arraybuffer_with_data, since these two functions already called the adjust_external_memory internal.

pub fn create_buffer_copy<D>(&self, data_to_copy: D) -> Result<JsBufferValue> where
D: AsRef<[u8]>,

This API allocates a node::Buffer object and initializes it with data copied from the passed-in buffer.

While this is still a fully-supported data structure, in most cases using a TypedArray will suffice.

pub fn create_arraybuffer(&self, length: usize) -> Result<JsArrayBufferValue>

pub fn create_arraybuffer_with_data(
&self,
data: Vec<u8>
) -> Result<JsArrayBufferValue>

pub unsafe fn create_arraybuffer_with_borrowed_data<Hint, Finalize>(
&self,
data: *const u8,
length: usize,
hint: Hint,
finalize_callback: Finalize
) -> Result<JsArrayBufferValue> where
Finalize: FnOnce(Hint, Env),

Safety

Mostly the same with create_arraybuffer_with_data

Provided finalize_callback will be called when Buffer got dropped.

You can pass in noop_finalize if you have nothing to do in finalize phase.

pub fn create_function(
&self,
name: &str,
callback: Callback
) -> Result<JsFunction>

This API allows an add-on author to create a function object in native code.

This is the primary mechanism to allow calling into the add-on’s native code from JavaScript.

The newly created function is not automatically visible from script after this call.

Instead, a property must be explicitly set on any object that is visible to JavaScript, in order for the function to be accessible from script.

pub fn create_function_from_closure<R, F>(
&self,
name: &str,
callback: F
) -> Result<JsFunction> where
F: 'static + Send + Sync + Fn(CallContext<'_>) -> Result<R>,
R: NapiValue,

pub fn get_last_error_info(&self) -> Result<ExtendedErrorInfo>

This API retrieves a napi_extended_error_info structure with information about the last error that occurred.

The content of the napi_extended_error_info returned is only valid up until an n-api function is called on the same env.

Do not rely on the content or format of any of the extended information as it is not subject to SemVer and may change at any time. It is intended only for logging purposes.

This API can be called even if there is a pending JavaScript exception.

pub fn throw_error(&self, msg: &str, code: Option<&str>) -> Result<()>

This API throws a JavaScript Error with the text provided.

pub fn throw_range_error(&self, msg: &str, code: Option<&str>) -> Result<()>

This API throws a JavaScript RangeError with the text provided.

pub fn throw_type_error(&self, msg: &str, code: Option<&str>) -> Result<()>

This API throws a JavaScript TypeError with the text provided.

pub fn fatal_error(self, location: &str, message: &str)

In the event of an unrecoverable error in a native module

A fatal error can be thrown to immediately terminate the process.

pub fn fatal_exception(&self, err: Error)

Trigger an ‘uncaughtException’ in JavaScript.

Useful if an async callback throws an exception with no way to recover.

pub fn define_class(
&self,
name: &str,
constructor_cb: Callback,
properties: &[Property<'_>]
) -> Result<JsFunction>

Create JavaScript class

pub fn wrap<T: 'static>(
&self,
js_object: &mut JsObject,
native_object: T
) -> Result<()>

pub fn unwrap<T: 'static>(&self, js_object: &JsObject) -> Result<&mut T>

pub fn unwrap_from_ref<T: 'static>(
&self,
js_ref: &Ref<()>
) -> Result<&'static mut T>

pub fn drop_wrapped<T: 'static>(&self, js_object: JsObject) -> Result<()>

pub fn create_reference<T>(&self, value: T) -> Result<Ref<()>> where
T: NapiValue,

This API create a new reference with the specified reference count to the Object passed in.

pub fn get_reference_value<T>(&self, reference: &Ref<()>) -> Result<T> where
T: NapiValue,

Get reference value from Ref with type check

Return error if the type of reference provided is mismatched with T

pub fn get_reference_value_unchecked<T>(&self, reference: &Ref<()>) -> Result<T> where
T: NapiValue,

Get reference value from Ref without type check

Using this API if you are sure the type of T is matched with provided Ref<()>.

If type mismatched, calling T::method would return Err.

pub fn create_external<T: 'static>(
&self,
native_object: T,
size_hint: Option<i64>
) -> Result<JsExternal>

If size_hint provided, Env::adjust_external_memory will be called under the hood.

If no size_hint provided, global garbage collections will be triggered less times than expected.

If getting the exact native_object size is difficult, you can provide an approximate value, it’s only effect to the GC.

pub fn get_value_external<T: 'static>(
&self,
js_external: &JsExternal
) -> Result<&mut T>

pub fn create_error(&self, e: Error) -> Result<JsObject>

pub fn spawn<T: 'static + Task>(&self, task: T) -> Result<AsyncWorkPromise<'_>>

Run Task in libuv thread pool, return AsyncWorkPromise

pub fn run_in_scope<T, F>(&self, executor: F) -> Result<T> where
F: FnOnce() -> Result<T>,

pub fn get_global(&self) -> Result<JsGlobal>

pub fn get_napi_version(&self) -> Result<u32>

pub fn get_uv_event_loop(&self) -> Result<*mut uv_loop_s>

pub fn add_env_cleanup_hook<T, F>(
&mut self,
cleanup_data: T,
cleanup_fn: F
) -> Result<CleanupEnvHook<T>> where
T: 'static,
F: 'static + FnOnce(T),

pub fn remove_env_cleanup_hook<T>(
&mut self,
hook: CleanupEnvHook<T>
) -> Result<()> where
T: 'static,

pub fn create_threadsafe_function<T: Send, V: NapiValue, R: 'static + Send + FnMut(ThreadSafeCallContext<T>) -> Result<Vec<V>>>(
&self,
func: &JsFunction,
max_queue_size: usize,
callback: R
) -> Result<ThreadsafeFunction<T>>

pub fn execute_tokio_future<T: 'static + Send, V: 'static + NapiValue, F: 'static + Send + Future<Output = Result<T>>, R: 'static + Send + Sync + FnOnce(&mut Env, T) -> Result<V>>(
&self,
fut: F,
resolver: R
) -> Result<JsObject>

pub fn create_date(&self, time: f64) -> Result<JsDate>

This API does not observe leap seconds; they are ignored, as ECMAScript aligns with POSIX time specification.

This API allocates a JavaScript Date object.

JavaScript Date objects are described in Section 20.3 of the ECMAScript Language Specification.

pub fn set_instance_data<T, Hint, F>(
&self,
native: T,
hint: Hint,
finalize_cb: F
) -> Result<()> where
T: 'static,
Hint: 'static,
F: FnOnce(FinalizeContext<T, Hint>),

This API associates data with the currently running Agent. data can later be retrieved using Env::get_instance_data().

Any existing data associated with the currently running Agent which was set by means of a previous call to Env::set_instance_data() will be overwritten.

If a finalize_cb was provided by the previous call, it will not be called.

pub fn get_instance_data<T>(&self) -> Result<Option<&'static mut T>> where
T: 'static,

This API retrieves data that was previously associated with the currently running Agent via Env::set_instance_data().

If no data is set, the call will succeed and data will be set to NULL.

pub fn strict_equals<A: NapiValue, B: NapiValue>(
&self,
a: A,
b: B
) -> Result<bool>

This API represents the invocation of the Strict Equality algorithm as defined in Section 7.2.14 of the ECMAScript Language Specification.

pub fn get_node_version(&self) -> Result<NodeVersion>

pub fn raw(&self) -> napi_env

get raw env ptr